Curve, PartialCurve: added a new directory to better structure the curve class
